Remove old data file, the-counted.csv, now unused.
When the data was only for 2015 there was only one data file,
the-counted.csv. Now data for 2016 is being collected too the data has
been split into two files, the-counted-2015.csv and
the-counted-2016.csv. The original data file no longer appears in the
Guardian's ZIP file (the data source) but it was still being kept here.
With this commit, it's gone.